requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.5
transformers>=4.41.0
torch>=2.2.0
python-pptx>=0.6.23
gradio>=4.36.0
tqdm>=4.66.0
pillow>=10.0.0
requests>=2.31.0
pytest>=8.0.0
